+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Update to `font8x8 = 0.2` in Cargo.toml.
 - Replace `FontUtf16` with `FontUnicode` type.
 - Replace `Utf16Fonts` with `UnicodeFonts` trait.
+- Replace `u16` with `char`.
+- Replace `FontString::encode_utf16` with `FontString::chars` method (breaking change).
 ## [v0.1.11] - 2018-06-17
 ### Fixed
 - URLs to `examples/scroll-*.rs` on README.